(#5) Introduce the example PandC which was earlier nCk
(PandC includes both nCk and nPk)
